feat: StudyApplication, StudyCancel 로 수정, url로 받는 것으로 수정
