+++ apps/client/app/(afterLogin)/study-apply/@modal/(.)study-cancel/[studyId]/page.tsx
@@ -5,16 +5,13 @@ import { Flex } from "@styled-system/jsx";
 import { Modal, Space, Text } from "@wow-class/ui";
 import { studyApplyApi } from "apis/studyApplyApi";
 import { tags } from "constants/tags";
-import { useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { revalidateTagByName } from "utils/revalidateTagByName";
 import Button from "wowds-ui/Button";
-const ApplyModal = () => {
-  const searchParams = useSearchParams();
 
-  const studyId = searchParams.get("studyId");
-
-  const [applySuccess, setApplySuccess] = useState(false);
+const StudyCancel = ({ params }: { params: { studyId: number } }) => {
+  const studyId = params.studyId;
+  const [cancelSucces, setCancelSuccess] = useState(false);
   const [studyTitle, setStudyTitle] = useState("");
 
   useEffect(() => {
@@ -34,33 +31,32 @@ const ApplyModal = () => {
     fetchStudyData();
   }, [studyId]);
 
-  const handleClickApplyButton = async () => {
-    const result = await studyApplyApi.applyStudy(Number(studyId));
+  const handleClickCancelButton = async () => {
+    const result = await studyApplyApi.cancelStudyApplication(Number(studyId));
+
     if (result.success) {
       revalidateTagByName(tags.studyApply);
-      setApplySuccess(true);
+      setCancelSuccess(true);
     }
   };
 
   return (
     <Modal>
       <Flex direction="column" textAlign="center" width="21rem">
-        {applySuccess ? (
+        {cancelSucces ? (
           <Text typo="h1">
             <span className={titleStyle}>{studyTitle}</span>
             <br />
-            신청이 완료되었어요.
+            수강이 취소되었어요.
           </Text>
         ) : (
           <>
             <Text typo="h1">
-              <span className={titleStyle}>{studyTitle}</span>을(를) <br />
-              신청하시겠습니까?
+              <span className={titleStyle}>{studyTitle}</span>의 <br />
+              수강을 취소하시겠습니까?
             </Text>
-            <Space height={22} />
-            <Text color="sub">한 번에 하나의 강의만 수강할 수 있어요.</Text>
-            <Space height={28} />
-            <Button onClick={handleClickApplyButton}>수강 신청하기</Button>
+            <Space height={38} />
+            <Button onClick={handleClickCancelButton}>신청 취소하기</Button>
           </>
         )}
       </Flex>
@@ -68,7 +64,7 @@ const ApplyModal = () => {
   );
 };
 
-export default ApplyModal;
+export default StudyCancel;
 
 const titleStyle = css({
   color: "primary",
