+++ apps/client/app/(afterLogin)/study-apply/study-application/[studyId]/page.tsx
@@ -1,8 +1,8 @@
 import { routePath } from "constants/routePath";
 import { redirect } from "next/navigation";
 
-const CancelModalPage = () => {
+const StudyApplicationPage = () => {
   return redirect(routePath["study-apply"]);
 };
 
-export default CancelModalPage;
+export default StudyApplicationPage;
