+++ apps/client/app/(afterLogin)/study-apply/study-cancel/[studyId]/page.tsx
@@ -1,8 +1,8 @@
 import { routePath } from "constants/routePath";
 import { redirect } from "next/navigation";
 
-const ApplyModalPage = () => {
+const StudyCancelPage = () => {
   return redirect(routePath["study-apply"]);
 };
 
-export default ApplyModalPage;
+export default StudyCancelPage;
