+++ apps/client/constants/routePath.ts
@@ -4,8 +4,8 @@ export const routePath = {
   ["my-study"]: "/my-study",
   ["my-assignment"]: "/my-study/my-assignment",
   ["study-apply"]: "/study-apply",
-  ["study-apply-modal"]: "/study-apply/application",
-  ["study-cancel-modal"]: "/study-apply/cancel",
+  ["study-apply-modal"]: "/study-apply/study-application",
+  ["study-cancel-modal"]: "/study-apply/study-cancel",
   ["auth-error-during-recruitment"]: "/auth-error-during-recruitment",
   ["auth-error-after-recruitment"]: "/auth-error-after-recruitment",
   ["attendance-check"]: "/my-study/attendance-check",
